+++ qc2/algorithms/base/qc2_qubit_mapper_base_class.py
